Updated precall screen to use custom hook
